backend/requirements.txt
fastapi==0.115.8
google-genai==1.7.0
python-dotenv==1.0.0
uvicorn==0.32.0
